Renames pipe -> createPipe
