+++ source/create-pipe.js
